Moved http_requests package outside of the common folder. Killed common.
