+++ ocean_lib/models/data_token.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 import requests
 from enforce_typing import enforce_types
 from eth_utils import remove_0x_prefix
-from ocean_lib.common.http_requests.requests_session import get_requests_session
+from ocean_lib.http_requests.requests_session import get_requests_session
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
 from ocean_lib.web3_internal.contract_base import ContractBase
 from ocean_lib.web3_internal.currency import from_wei, pretty_ether_and_wei, to_wei
